refactor: factor out request_pool structure
