+++ execution-engine/contracts/hdac-system/pop/src/duration_queue.rs
@@ -1,42 +1,99 @@
-mod duration_queue;
-mod requests;
+use alloc::{boxed::Box, vec::Vec};
+use core::result;
 
-use alloc::vec::Vec;
+use types::{
+    bytesrepr::{self, FromBytes, ToBytes, U64_SERIALIZED_LENGTH},
+    system_contract_errors::pos::{Error, Result},
+    BlockTime, CLType, CLTyped,
+};
 
-use contract::contract_api::storage;
+#[derive(Default)]
+pub struct DurationQueue<T: DurationQueueItem>(pub Vec<DurationQueueEntry<T>>);
 
-use crate::constants::local_keys;
-
-use duration_queue::DurationQueue;
-pub use requests::{ClaimRequest, RedelegateRequest, UndelegateRequest};
-
-pub fn read_undelegation_requests() -> DurationQueue<UndelegateRequest> {
-    storage::read_local(&local_keys::UNDELEGATE_REQUEST_QUEUE)
-        .unwrap_or_default()
-        .unwrap_or_default()
+#[derive(Debug, Clone, Copy)]
+pub struct DurationQueueEntry<T: DurationQueueItem> {
+    pub item: T,
+    pub timestamp: BlockTime,
 }
 
-pub fn read_redelegation_requests() -> DurationQueue<RedelegateRequest> {
-    storage::read_local(&local_keys::REDELEGATE_REQUEST_QUEUE)
-        .unwrap_or_default()
-        .unwrap_or_default()
+pub trait DurationQueueItem: CLTyped + FromBytes + ToBytes + Copy + Clone {}
+
+impl<T: DurationQueueItem> DurationQueue<T> {
+    pub fn push(&mut self, item: T, timestamp: BlockTime) -> Result<()> {
+        if let Some(entry) = self.0.last() {
+            if entry.timestamp > timestamp {
+                return Err(Error::TimeWentBackwards);
+            }
+        }
+        self.0.push(DurationQueueEntry { item, timestamp });
+        Ok(())
+    }
+    pub fn pop_due(&mut self, timestamp: BlockTime) -> Vec<DurationQueueEntry<T>> {
+        let (older_than, rest) = self
+            .0
+            .iter()
+            .partition(|entry| entry.timestamp <= timestamp);
+        self.0 = rest;
+        older_than
+    }
 }
 
-pub fn write_undelegation_requests(queue: DurationQueue<UndelegateRequest>) {
-    storage::write_local(local_keys::UNDELEGATE_REQUEST_QUEUE, queue);
+impl<T: DurationQueueItem> FromBytes for DurationQueue<T> {
+    fn from_bytes(bytes: &[u8]) -> result::Result<(Self, &[u8]), bytesrepr::Error> {
+        let (len, mut bytes) = u64::from_bytes(bytes)?;
+        let mut queue = Vec::new();
+        for _ in 0..len {
+            let (entry, rest) = DurationQueueEntry::from_bytes(bytes)?;
+            bytes = rest;
+            queue.push(entry);
+        }
+        Ok((DurationQueue(queue), bytes))
+    }
 }
 
-pub fn write_redelegation_requests(queue: DurationQueue<RedelegateRequest>) {
-    storage::write_local(local_keys::REDELEGATE_REQUEST_QUEUE, queue);
+impl<T: DurationQueueItem> ToBytes for DurationQueue<T> {
+    fn to_bytes(&self) -> result::Result<Vec<u8>, bytesrepr::Error> {
+        let mut bytes = (self.0.len() as u64).to_bytes()?; // TODO: Allocate correct capacity.
+        for entry in &self.0 {
+            bytes.append(&mut entry.to_bytes()?);
+        }
+        Ok(bytes)
+    }
+    fn serialized_length(&self) -> usize {
+        U64_SERIALIZED_LENGTH + self.0.iter().map(ToBytes::serialized_length).sum::<usize>()
+    }
 }
 
-pub fn read_claim_requests() -> Vec<ClaimRequest> {
-    storage::read_local(&local_keys::CLAIM_REQUESTS)
-        .unwrap_or_default()
-        .unwrap_or_default()
+impl<T: DurationQueueItem> CLTyped for DurationQueue<T> {
+    fn cl_type() -> CLType {
+        CLType::List(Box::new(DurationQueueEntry::<T>::cl_type()))
+    }
 }
-pub fn write_claim_requests(list: Vec<ClaimRequest>) {
-    storage::write_local(local_keys::CLAIM_REQUESTS, list);
+
+impl<T: DurationQueueItem> FromBytes for DurationQueueEntry<T> {
+    fn from_bytes(bytes: &[u8]) -> result::Result<(Self, &[u8]), bytesrepr::Error> {
+        let (item, bytes) = T::from_bytes(bytes)?;
+        let (timestamp, bytes) = BlockTime::from_bytes(bytes)?;
+        let entry = DurationQueueEntry { item, timestamp };
+        Ok((entry, bytes))
+    }
+}
+
+impl<T: DurationQueueItem> ToBytes for DurationQueueEntry<T> {
+    fn to_bytes(&self) -> result::Result<Vec<u8>, bytesrepr::Error> {
+        Ok((self.item.to_bytes()?.into_iter())
+            .chain(self.timestamp.to_bytes()?)
+            .collect())
+    }
+    fn serialized_length(&self) -> usize {
+        self.item.serialized_length() + self.timestamp.serialized_length()
+    }
+}
+
+impl<T: DurationQueueItem> CLTyped for DurationQueueEntry<T> {
+    fn cl_type() -> CLType {
+        CLType::Any
+    }
 }
 
 #[cfg(test)]
@@ -45,7 +102,9 @@ mod tests {
 
     use types::{account::PublicKey, system_contract_errors::pos::Error, BlockTime, U512};
 
-    use super::{delegation_queue::DurationQueueEntry, DurationQueue, UndelegateRequest};
+    use crate::local_store::UndelegateRequest;
+
+    use super::{DurationQueue, DurationQueueEntry};
 
     const KEY1: [u8; 32] = [1; 32];
     const KEY2: [u8; 32] = [2; 32];
