+++ execution-engine/contracts/hdac-system/pop/src/lib.rs
@@ -3,6 +3,8 @@
 extern crate alloc;
 
 mod constants;
+mod duration_queue;
+mod local_store;
 mod math;
 mod pop_contract;
